+++ backend/src/models/asset.models.js
@@ -4,7 +4,7 @@ const AssetSchema = new mongoose.Schema({
     _id: { type: String, required: true },
     id: { type: Number, required: true, unique: true },
     name: { type: String, required: true },
-    is_deleted: { type: Boolean, default: false }
+    isDeleted: { type: Boolean, default: false }
 }, { versionKey: false });
 
 export default mongoose.model('assets', AssetSchema);