+++ frontend/src/api/auth.ts
@@ -5,11 +5,6 @@ export type Credentials = {
     password: string;
 };
 
-export async function register(credentials: Credentials) {
-    const res = await getApi().post('/api/auth/register', credentials);
-    return res.data['token'];
-};
-
 export async function login(credentials: Credentials) {
     const res = await getApi().post('/api/auth/login', credentials);
     return res.data['token'];